# Simplify interaction between Chronometer and FTL