--- a/advanced/Scripts/chronometer.sh
+++ b/advanced/Scripts/chronometer.sh
@@ -14,7 +14,7 @@ pihole-FTL() {
   if [[ -n "$ftl_port" ]]; then
     # Get data from FTL
     C=0
-    readarray -t LINE <<< "$(echo ">$1" | nc 127.0.0.1 "${ftl_port}")"
+    readarray -t LINE <<< "$(echo ">$1" | nc -w 1 127.0.0.1 "${ftl_port}")"
     until [[ "${LINE[$C]}" == *"EOM"* || "${LINE[$C]}" == "" ]]; do
       echo "${LINE[$C]}"
       let C+=1