@WaLLy3K 2017-06-21
Minimum viable code test:
```
#!/usr/bin/env bash
calc(){ awk "BEGIN {print $*}" 2> /dev/null; }

pihole-FTL-Old() {
  ftl_port=$(cat /var/run/pihole-FTL.port 2> /dev/null)
  if [[ -n "$ftl_port" ]]; then
    # Open connection to FTL
    exec 3<>"/dev/tcp/localhost/$ftl_port"

    # Test if connection is open
    if { "true" >&3; } 2> /dev/null; then
      # Send command to FTL
      echo -e ">$1" >&3

      # Read input
      read -r -t 1 LINE <&3
      until [[ ! $? ]] || [[ "$LINE" == *"EOM"* ]]; do
         echo "$LINE" >&1
         read -r -t 1 LINE <&3
      done

      # Close connection
      exec 3>&-
      exec 3<&-
   fi
  else
    echo "FTL offline"
  fi
}

pihole-FTL-New() {
  ftl_port=$(cat /var/run/pihole-FTL.port 2> /dev/null)
  if [[ -n "$ftl_port" ]]; then
    # Get data from FTL
    C=0
    readarray -t LINE <<< "$(echo ">$1" | nc -w 1 127.0.0.1 "${ftl_port}")"
    if [[ "${#LINE[@]}" -ge "1" && ! "${LINE[0]}" == "" ]]; then
      until [[ "${LINE[$C]}" == *"EOM"* || "${LINE[$C]}" == "" ]]; do
        echo "${LINE[$C]}"
        let C+=1
      done
    else
      echo -e "FTL offline"
    fi
  else
    echo -e "FTL offline"
  fi
}

run_start=$(date +%s.%N)
printf "Start: %.2fs\n" $(calc "$(date +%s.%N) - $run_start")
pihole-FTL-Old "stats"
printf "Runtime for old FTL code: %.2fs\n" $(calc "$(date +%s.%N) - $run_start")

run_start=$(date +%s.%N)
pihole-FTL-New "stats"
printf "Runtime for new FTL code: %.2fs\n" $(calc "$(date +%s.%N) - $run_start")
```

Output returns:
```
>stats output
Runtime for old FTL code: 0.02s
>stats output
Runtime for new FTL code: 2.03s
```

That's taking way too long to generate - are you seeing the same run times on your end?