@DL6ER 2017-06-21
```
pi@raspberrypi:~ $ time pihole-FTL-Old version
version v2.9.1-9-g60bf275
tag v2.9.1
branch development
date 2017-06-19 20:10:38 +0200

real	0m0.252s
user	0m0.000s
sys	0m0.000s

pi@raspberrypi:~ $ time pihole-FTL-New version
version v2.9.1-9-g60bf275
tag v2.9.1
branch development
date 2017-06-19 20:10:38 +0200

real	0m0.521s
user	0m0.000s
sys	0m0.010s
```