@DL6ER 2017-06-21
A more scientific approach: I ran the timing shown above 25 times for either of the two commands and computed the statistics. I see:

## New
Average: 0.28s +/- 0.04s
Max: 0.53s
Min: 0.24s

## Old
Average: 0.31s +/- 0.10s
Max: 0.74s
Min: 0.23s