@WaLLy3K 2017-06-22
I'm just not seeing that kind of swiftness when I try and run the proposed code, which is a bit confusing:

```
#!/usr/bin/env bash
calc(){ awk "BEGIN {print $*}" 2> /dev/null; }

ftlFunc() {
  ftl_port=$(cat /var/run/pihole-FTL.port 2> /dev/null)
  if [[ -n "$ftl_port" ]]; then
    # Get data from FTL
    C=0
    readarray -t LINE <<< "$(echo ">$1" | nc -w 1 127.0.0.1 "${ftl_port}")"
    if [[ "${#LINE[@]}" -ge "1" && ! "${LINE[0]}" == "" ]]; then
      until [[ "${LINE[$C]}" == *"EOM"* || "${LINE[$C]}" == "" ]]; do
        echo "${LINE[$C]}"
        let C+=1
      done
    else
      echo -e "FTL offline"
    fi
  else
    echo -e "FTL offline"
  fi
}

run_start=$(date +%s.%N)
ftlFunc version
printf "Runtime for new FTL code: %.2fs\n" $(calc "$(date +%s.%N) - $run_start")
```

```
version v2.9.3
tag v2.9.3
branch (no branch)
date 2017-06-21 20:19:49 +0200
Runtime for new FTL code: 2.02s
```