@DL6ER 2017-06-24
@WaLLy3K Any updats on this? Have you tried running the code the same way I did (define functions, run them through `time` instead of measuring the time like you did)?