@DL6ER 2017-06-24
Going to merge `development` now. Hopefully that resolves the test issues.