@WaLLy3K 2017-06-25
Running `time ftlFunc version`:
```
version v2.9.4
tag v2.9.4
branch (no branch)
date 2017-06-22 23:33:48 +0200

real    0m2.016s
user    0m0.000s
sys     0m0.000s
```

Old code completes in 0m0.008s.