@WaLLy3K 2017-06-29
Unless there's something *really* wrong with my setup (hey, always a possibility!), I'd consider closing this until we can find code that runs (almost) as quickly as what we already have.

As far as Chrono and my personal opinion goes, I absolutely do not mind that the code is *longer*, I just care that it's *faster*. People could be setting a Chrono refresh of 1 second (when as-is with #1518, the first-run of the code is roughly 0.5-0.6s), and we need to be able to retrieve data from `FTL` as quick as possible.